+++ pkg/controller/direct/securesourcemanager/mapper.generated.go
@@ -1,4 +1,4 @@
-// Copyright 2024 Google LLC
+// Copyright 2025 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -26,13 +26,35 @@ func Instance_HostConfig_FromProto(mapCtx *direct.MapContext, in *pb.Instance_Ho
 		return nil
 	}
 	out := &krm.Instance_HostConfig{}
+	// MISSING: HTML
+	// MISSING: Api
+	// MISSING: GitHTTP
+	// MISSING: GitSSH
+	return out
+}
+func Instance_HostConfig_ToProto(mapCtx *direct.MapContext, in *krm.Instance_HostConfig) *pb.Instance_HostConfig {
+	if in == nil {
+		return nil
+	}
+	out := &pb.Instance_HostConfig{}
+	// MISSING: HTML
+	// MISSING: Api
+	// MISSING: GitHTTP
+	// MISSING: GitSSH
+	return out
+}
+func Instance_HostConfigObservedState_FromProto(mapCtx *direct.MapContext, in *pb.Instance_HostConfig) *krm.Instance_HostConfigObservedState {
+	if in == nil {
+		return nil
+	}
+	out := &krm.Instance_HostConfigObservedState{}
 	out.HTML = direct.LazyPtr(in.GetHtml())
 	out.Api = direct.LazyPtr(in.GetApi())
 	out.GitHTTP = direct.LazyPtr(in.GetGitHttp())
 	out.GitSSH = direct.LazyPtr(in.GetGitSsh())
 	return out
 }
-func Instance_HostConfig_ToProto(mapCtx *direct.MapContext, in *krm.Instance_HostConfig) *pb.Instance_HostConfig {
+func Instance_HostConfigObservedState_ToProto(mapCtx *direct.MapContext, in *krm.Instance_HostConfigObservedState) *pb.Instance_HostConfig {
 	if in == nil {
 		return nil
 	}
@@ -41,6 +63,46 @@ func Instance_HostConfig_ToProto(mapCtx *direct.MapContext, in *krm.Instance_Hos
 	out.Api = direct.ValueOf(in.Api)
 	out.GitHttp = direct.ValueOf(in.GitHTTP)
 	out.GitSsh = direct.ValueOf(in.GitSSH)
+	return out
+}
+func Instance_PrivateConfig_FromProto(mapCtx *direct.MapContext, in *pb.Instance_PrivateConfig) *krm.Instance_PrivateConfig {
+	if in == nil {
+		return nil
+	}
+	out := &krm.Instance_PrivateConfig{}
+	out.IsPrivate = direct.LazyPtr(in.GetIsPrivate())
+	if in.GetCaPool() != "" {
+		out.CaPoolRef = &refs.PrivateCACAPoolRef{External: in.GetCaPool()}
+	}
+	return out
+}
+func Instance_PrivateConfig_ToProto(mapCtx *direct.MapContext, in *krm.Instance_PrivateConfig) *pb.Instance_PrivateConfig {
+	if in == nil {
+		return nil
+	}
+	out := &pb.Instance_PrivateConfig{}
+	out.IsPrivate = direct.ValueOf(in.IsPrivate)
+	if in.CaPoolRef != nil {
+		out.CaPool = in.CaPoolRef.External
+	}
+	return out
+}
+func Instance_PrivateConfigObservedState_FromProto(mapCtx *direct.MapContext, in *pb.Instance_PrivateConfig) *krm.Instance_PrivateConfigObservedState {
+	if in == nil {
+		return nil
+	}
+	out := &krm.Instance_PrivateConfigObservedState{}
+	out.HTTPServiceAttachment = direct.LazyPtr(in.GetHttpServiceAttachment())
+	out.SSHServiceAttachment = direct.LazyPtr(in.GetSshServiceAttachment())
+	return out
+}
+func Instance_PrivateConfigObservedState_ToProto(mapCtx *direct.MapContext, in *krm.Instance_PrivateConfigObservedState) *pb.Instance_PrivateConfig {
+	if in == nil {
+		return nil
+	}
+	out := &pb.Instance_PrivateConfig{}
+	out.HttpServiceAttachment = direct.ValueOf(in.HTTPServiceAttachment)
+	out.SshServiceAttachment = direct.ValueOf(in.SSHServiceAttachment)
 	return out
 }
 func Repository_InitialConfig_FromProto(mapCtx *direct.MapContext, in *pb.Repository_InitialConfig) *krm.Repository_InitialConfig {
@@ -96,7 +158,7 @@ func SecureSourceManagerInstanceObservedState_FromProto(mapCtx *direct.MapContex
 	// MISSING: Labels
 	out.State = direct.Enum_FromProto(mapCtx, in.GetState())
 	out.StateNote = direct.Enum_FromProto(mapCtx, in.GetStateNote())
-	out.HostConfig = Instance_HostConfig_FromProto(mapCtx, in.GetHostConfig())
+	out.HostConfig = Instance_HostConfigObservedState_FromProto(mapCtx, in.GetHostConfig())
 	return out
 }
 func SecureSourceManagerInstanceObservedState_ToProto(mapCtx *direct.MapContext, in *krm.SecureSourceManagerInstanceObservedState) *pb.Instance {
@@ -110,7 +172,7 @@ func SecureSourceManagerInstanceObservedState_ToProto(mapCtx *direct.MapContext,
 	// MISSING: Labels
 	out.State = direct.Enum_ToProto[pb.Instance_State](mapCtx, in.State)
 	out.StateNote = direct.Enum_ToProto[pb.Instance_StateNote](mapCtx, in.StateNote)
-	out.HostConfig = Instance_HostConfig_ToProto(mapCtx, in.HostConfig)
+	out.HostConfig = Instance_HostConfigObservedState_ToProto(mapCtx, in.HostConfig)
 	return out
 }
 func SecureSourceManagerInstanceSpec_FromProto(mapCtx *direct.MapContext, in *pb.Instance) *krm.SecureSourceManagerInstanceSpec {
